feat: support "hoistUseStatements" loader option
https://github.com/shakacode/sass-resources-loader v2.1.0 added support
for the "hoistUseStatements" option. Add support for it in the module.

Implemented with a least resistance by adding it as a top-level option
without ability to specify it individually for each resource type.
If someone requests that later, we could add support for it by accepting
an array of resource types.